Update rationality_colorwheel.pde
I fixed a bug. The original colorwheel contains errors for its upper and lower bounds. When, for example, that version calculated the lower bound for the Unlikely probability region it used 0.6 rather than 0.5. It turns out that this subtle difference matters. For instance, the lower bound for Orange Green Orange is Orange, but using the 0.6 approximation we get Yellow.